backend/requirements.txt
azure-ai-projects==1.0.0
azure-ai-voicelive[aiohttp]>=1.0.0
azure-storage-blob>=12.20.0
azure-cognitiveservices-speech==1.47.0
azure-identity>=1.25.1
azure-monitor-opentelemetry>=1.6.0
alembic>=1.16.0
flask==3.1.2
flask-sock==0.7.0
github-copilot-sdk
openai==2.13.0
pillow>=11.2.1
psycopg[binary]>=3.2.0
python-dotenv==1.2.1
pyyaml==6.0.3
SQLAlchemy>=2.0.0
websockets==15.0.1